vmusb readout: set NIM O2 active while in DAQ mode
As vmusb does not allow to directly output the DAQ state on one of the
two NIM outputs the current solution uses the "USB trigger" bit of the
action register together with keeping the NIM output latched during DAQ
mode.

This change also uses the top yellow LED to signal "USB InFIFO Full".