improve timetick handling
During fast DAQ readouts where the analysis could not
keep up with the data rate the timetick count was misleading
as timeticks where transported through the buffer queue and could
thus be dropped just like data buffers. To fix this timeticks
are now generated on the analysis side during readout.

For replays the timeticks from the replay are used as before.