vme readout templates: increase MBLT transfer count
The transfer count is now the minimum of the max supported
transfer count of the controllers. This is done to not run
into the case where large readout data is cut off just because
not enough block transfer where specified in the readout script.

The following transfer count limits apply:
         BLT                   MBLT
VMUSB    8388608 (1u<<23)      4194302 (1u<<22)
SIS3153  4194303 (0xffffff/4)  2097151 (0xffffff/8)