change how analysis modifications are handled
Instead of rebuilding the analysis objects immediately after modifying the
internal structures, the system now keeps track of which objects need to be
rebuilt. Then in Analysis::beginRun() the objects that have been marked
will have their beginRun() invoked.

This means beginRun() will now only be called if the operator
was modified or the RunInfo or VMEMap changed. To be able to still keep
or clear histogram state a new clearState() method has been added to
PipeSourceInterface and reimplemented in the sink classes that can be
cleared. This will be called from Analysis::beginRun() if needed.