SIS3153 readout: activate OUT2 during main readout stack execution
OUT2 is now set to logical 1 during execution of the main readout stack.
This main stack is currently auto assigned to the first non-timer,
non-watchdog event.

During setup the stack is marked with "OUT2" in the log messages.

Tests where done with an MTDC-32 in single readout mode using the modules
internal pulser. With this setup the MTDC generates data faster than the
readout can handle which means a new interrupt will immediately trigger
after execution of the readout stack.

The following signal times where measured on the output OUT2 in NIM
configuration:
Logical 1: 1.4 us
Logical 0 in between executions of the stack: 0.7 us