disable VMUSB NIM output activation code
The outputs do not seem to behave according to the documentation. Also
the exact meaning of the latch bit is still unclear.

We thus decided to not setup the NIM outputs during the DAQ start sequence
anymore. Instead VME script commands to write certain specific registers
will be implemented to enable the user to configure the outputs as desired.