remove prototype threading code from a2
Reasons: Doing operator work in threads was not efficient. The individual
work units (one operator step) where too small to justify the overhead
caused by having to queue up work. The effect was that more CPU cores where
utilized but there was no improvement in overall throughput.

Also with the introduction of conditions this way of threading would have
required locking of the A2 global bitfield which would have further slowed
down the system.

The next thing to try out is thread histogramming. See the comment at the
bottom of a2.cc for some initial thoughts about this.