more work on condition ui
(De)Applying condition links is almost working now. There are still bugs
with the modification buttons sometimes not being visible or staying
visible.

It would also be good to have the condition tree get "node locked" if
modifications are pending. This would force the user to either accept or
reject the modifications before moving on to another condition.