fix race condition during DAQ/replay start
The race was between the stream worker thread rebuilding the analysis and
the main thread using the analysis structures to update UI elements.
This change moves rebuilding the analysis into the main thread which should
solve the problem.
Testing showed that even big analysis setups have a build time of <0.1s so
blocking the GUI should not be an issue.

# Conflicts:
#	src/mvme_context.cc