fixes for the ROOT treewriter client
Replace illegal characters in branch names with underscores. This fixes
crashes related to branch names.

Add missing call to TFile::Write() at the end right before closing the
file. Without this call not all data tree data was actually written to the
file and in the case of very short runs the ROOT file did not have an index
at all.