handle connection loss if detected while running a vme script
This is handled centrally in MVMEContext::runScript. If the connection to
the vme controller is lost as indicated by the return values from running
the vme script then the VMEController instance is manually disconnected
to correctly reflect the new state.

This only works for MVLC USB and other controllers which can signal a
connection loss. For MVLC ETH connection loss can often not be detected
because the only indication is that the reads are timing out. To handle
this there would need to be some logic that checks if a certain number of
reads timed out in a row and then assume that the connection is lost.