hack in a way to process fixed register readouts
Module readouts consisting of only a prefix, meaning they contain a
fixed amount of single register reads and nothing else are now passed to
the analysis as if the readout yielded a dynamic block.

This is done because there's no way to handle prefix/suffix data in the
current analysis but the MVLC parser produces these values.