mvlc usb: limit the time spent waiting for a clean readout shutdown
Prior to this commit the logic used to shutdown a readout could hang
forever if disabling the mvlc triggers did not work for some reason and
the mvlc continued to produce data.

Now the time spent in the loop is limited to 10s. This means the
application will at least return from the call and continue to work altough
the mvlc will probably be in a bad state at that point.