fix vme controller related ui hangs part1
Move the ctrl->isOpen() check from the gui thread into the pooled thread
which attempts to open the controller.

This way when using the MVLC the gui thread will not hit the mvlc
isConnected() method and thus won't have to take both of the pipe locks
which is what causes the stall.