fix possible stale data when using the MVLC
Getting stale buffers was a side effect from the change to stop the analyis
immediately instead of waiting for it to process all buffers. So if the
analysis could not keep up and the DAQ was stopped there would still be
filled buffers in the queue which would be consumed by the analyis on
the next run.