fix sluggish gui with mvlc_eth at low data rate
The fix involves using a separate mutex to guard the per-pipe and
per-channel stats of the mvlc_eth implementation instead of reusing
the normal pipe mutex. By having a separate mutex the stats structures can
be retrieved during a read operation and this operation is decoupled from
the pipes read timeout.