limit the number of attempts when reading a MVLC command response
This works around an issue where the MVLC DAQ could not be stopped:
- MVLC produced timeout error notifications at a high rate
- DAQ should be stopped by disabling the triggers for each stack
- No response was received as a result of disabling the triggers for
  stack0. Instead only error notifications where received. The loop never
  exited and thus mvme was stuck in the DAQ stop procedure.
- Also the notification buffer was growing indefinitely.

Note: this still doesn't actually stop the DAQ because
disable_all_triggers() will return an error trying to disable the first
stacks triggers. It will not continue with the next stacks.

This needs a better fix. Maybe a fire and forget type of the writeRegister
method or a firmware change to limit the number of notifications.