update to mdpp16 to mdpp16_scp config update
Instead of using startsWith("mdpp16") the test is now contains("mdpp16")
an uses a case insensitive comparison. This should catch more cases.

It still save to do this even if the setup contains mdpp16_rcp and
mdpp16_qdc because these modules will never lead to the generation of an
empty module type string in the vme config json.