do not raise vme script editors on vme tree single click
Raising an open vme script editor windows on single click interfered with
keyboard shortcuts because the focus was immediately taken away from the
tree widget. So pressing ctrl-c after single-clicking a vme script node
was always performing a copy operation in the text editor for the vme
script.

The same issue also lead to F2 never working to rename vme scripts that
where open in an editor.

Now vme script nodes have to be double clicked for their editor window to
be raised to the front.