fix vme config version not being updated when saving an existing vme config
The version information in the json was not updated before writing out
the config file. Thus when next loading the config the schema update
procedure was run again resetting any previously updated event variables
to their default values (because guessing the correct values was not
possible anymore from the already updated scripts).
This was not an issue when creating a new config as it always contained the
most current version.