fix analysis Directory objects not being updated by auto-assign on load
When loading an analysis where the object ids do not match the vme config
object ids an auto-assign step tries to correctly match the analysis
to the ones in the vme config and then updates the analysis (rewrite).

When adding Directory objects to the analysis I did not add code to rewrite
these on load. This resulted in the directories not showing up in the
analysis UI. The contained objects where still created and correctly
connected.