speed up the 'save vme config' action
Instead of walking the config hierarchy and resetting the 'modified' state
of every child object now only the root config object has its state
reset.

This in turn makes the analysis display repopulate itself only once when
saving which can be a significant benefit when using a setup with many
modules/events.