[vme_script] drop support for the counted block read commands
These commands first read the cycle count from a register and then
performed the (M)BLT block transfer for that many cycles.

VMUSB and SIS support it but with different options, the MVLC does not
support it. As long as a module issues BERR or can handle always reading
a fixed number of cycles there should be no need for these special
commands.