cleanup cmake warning and optimization flags
Debug and Optimization flags are not touched anymore. More warnings are
enabled via -Wall and -Wextra, c++14 is enabled via cmake, not by manually
adding the flag to CMAKE_CXX_FLAGS.

This needs testing for all platforms and the debug/release flags have to
be set on the buildbot system for automatic builds.