[mvlc-refactor][wip] switch to mesytec-mvlc lib implementation
This commit removes a lot of now duplicate code and switches callers to
the new library implementation.

Readout, replays and the analysis side still need to be ported to the
abstractions provided by the library.

Also the MVLCObject and MVLC_VMEController implementations need testing,
although quick tests via the GUI seemed ok.