port the mvlc stream/analysis side to the new mvlc lib
This implementation is mostly working. Setups with multiple modules and
events as well as setups using multievent readouts need testing.

Also the issue where the analysis info/debug widget is not updating
properly during a run seems to be more prevalent now. In some cases the
analysis side did not seem to receive any data at all. This might be caused
by non-threadsafe access to the analysis or a startup race condition.
Figure this out!