track down and fix an issue where the listfile directory was overwritten
The compression selection combo box emits an initial currentIndexChanged()
which leads to the DAQControlWidget emitting listfileOutputInfoModified()
which made MVMEContext update the mvmeworkspace.ini file. All this happened
before the loaded ListfileOutputInfo was set on the DAQControl widget.

The whole interaction and updating is bad. Maybe it's better to pull the
current compression setting on daq start and/or on exit?