add DAQ run notes feature
Run notes are kept in a text file in the workspace directory. When a DAQ
run is stopped the current contents of the file are added to the listfile
archive. A simple editor GUI can be opened via a new button added to the
DAQControl widget.

When replaying from file the run notes widget is populated with the
contents from the listfile archive and is set to read-only.