improve the vme config json schema update code slightly
Adds a version of add_event_variables() which does not create internal
EventConfig and other objects but instead operates on the JSON data
directly. This allows making changes to the EventConfig class as it does
not need to be able to deserialize from old JSON data anymore.

Note: this also needs to be done for update_event_scripts() and
update_module_scripts()!

Also backs out the v4_to_v5 conversion as we went back to v4