[mvlc] Deactivate StackStart0 in the default setup
Previously the StackStart unit was set to active and intended to trigger
the periodic readout for the counter readout event. If the user doesn't
create that second periodic event the StackStart unit would try to run
stack #2 which is not setup and thus defaults to executing the same
commands as stack #0, the direct command stack. This would effectively
make setups with the default script and no added periodic event rerun the
last direct VME commands once every second.