vme_templates: add "stop acq" sequence to the VME interface scripts
This will make it so that after initialization the module does not accept
triggers and does not yet generate VME IRQs.

This change is needed for the new MVLC start sequence:
- global daq start scripts
- init trigger io
- init modules			   <- modules would have produced data here
- upload readout stacks
- enable stack triggers    <- stacks would be triggered by the module
- mcast daq start          <- the multicast module "reset and go"
- trigger_io daq_start     <- trigger_io daq start signal

With this change in place the module can't trigger readouts stacks
prior to the mcast daq start script having been executed.