[vme_templates] improve mvlc_timestamper scripts
The VME Interface Settings script does not set the mcst of the MVLC
anymore. Reason: the counter is almost always used in a secondary periodic
event wich did overwrite the mcst of the MVLC specified in the primary
event. As a result the MVLC Timestamper was not reset anymore by the
primary events readout reset script.