create vme/analysis config files when opening from listfile
The idea is to have a copy of the config data from the listfiles on disk
and reload that copy on next opening the workspace. This way no vme vs
analysis conflicts will arise after closing listfiles (unless the user
takes specific action to create conflicts). The temporary filenames on disk
are not visible to the user and treated as empty filenames in the gui.