[vme_templates] default_filters: prefix source and operator names with the module name
Added a python script to automatically prepend 'module_name.' to sources
and operators in the default_filters.analysis file.

When instantiating the filters the module_name will be replaced by the
concrete module instances name. This makes is possible to distinguish where
e.g. an 'amplitude' analysis array originates from.