vme_tree: allow moving module between events via drag&drop
A check in the drop code prohibited moving modules from event to event.
Only the module order inside an event could be changed.
Now the module can be moved to another event. After the move the analysis
is paused and all objects that are dependents of the module are updated to
run in the new events execution context.