[vme_templates] stop acquisition in module reset scripts
Fix for an issue where mesytec modules could signal a VME IRQ during
initialization if the pulser was active or the module received triggers.

The reset scripts now disable the module DAQ, reset the FIFO and write
to the readout_reset register. This means when later on setting the IRQ
value, no IRQ can be immediately signalled as the module DAQ will not
be active.