analysis: suppress events where all module data is empty (SIS3153)
This should not be needed for correctly configured and working readouts as
at least one module per event should have readout data ready, otherwise
the readout should not have been triggered at all.

The emtpy events are still counted in the totalEvents counter but they
are not passed to the analysis at all. Massively speeds up replays when
they contain a lot of these empty events.