mvlc stream worker: avoid deadlocks when accessing multi event splitter counters
When the analysis side was paused the stream worker held the counters lock.
At the same time the GUI was trying to update the counters display.

Now a protected copy of the internal splitter counters is used to provide
thread safe counter access to clients. Parser and splitter counters are
now updated at the end of processBuffer().