change file saving code across mvme and most gui dev tools
The logic where QFileDialog::getSaveFileName() was used and a missing
filename extension was added later on was flawed: existing files could be
overwritten without the users permission. Now a QFileDialog instance is
created and QFileDialog::setDefaultSuffix() is used. This works under
Windows and Linux with the exception of a Qt bug where a '.' in a parent
directory name prevents QFileDialog from adding the file extension itself.
https://bugreports.qt.io/browse/QTBUG-59401
More code is needed to work around that problem.