save and restore node expansion state of the vme tree
A new QSettings INI file called "vme_tree_ui_state.ini" is created in the
workspace. If a ConfigObject node is expanded the state is recorded in the
file. If the node is collapsed the state is removed from the file.

Limitations: the expansion state of the "Event/Readout Loop" and
"Event/Multicast DAQ Start/Stop" nodes cannot be recorded because they are
not associated with a ConfigObject in the VMEConfig.