vme_script: fix issue with lowercasing variable names at the start of lines
The first part of a parsed line (usually the command) is lowercased to
simplify command lookups. This lead to an issue of failed variable lookups
if the variable was used at the start of the line.

Lowercasing has now been moved out of the pre-parse step to right before
evaluating the command.