doc: update to the manual ARP setup instructions in the installation chapter
Background: with recent firmware versions the MVLC does not leave the DHCP
phase and accept the used destination IP-address as its own address when
receiving a ping request. This had to be changed because some DHCP servers
use ping requests to make some preliminary checks before handing out a
lease. With the old behavior this could lead to DHCP "not working" because
the MVLC accepted the address used in the prelimiary ping.