analysis: keep IntervalCondition data intact for 'ignored' intervals
The min/max values are now kept as is and the 'ignore' state is kept in a
separate bool variable. The a2 runtime has not changed and still interprets
invalid intervals as being ignored. The a2_adapter layer and the
serialization methods have been updated to deal with the changes.

Also when editing values in the table the x1/x2 cells are not swapped
anymore so that x1<2. Instead this interval normalization is only done
in the interval editor picker when rendering intervals.