mvme_jsonrpc_replay: hack fix for race between EventServer client and replay start
When loading a replay all EventServer clients get disconnected and need
time to reconnect. If the replay is started during that time the clients
will only receive partial data. Bad design...

The workaround is to sleep for 1s between loading the listfiles and
starting the replay. This way clients have time to reconnect.