listfile_filtering: fix outputting the wrong event
The next event was written to the output instead of the one that was used
to produce the condition results. The bug was hidden in the call
structure of the stream worker and the analysis. Hacky fix was to add
module data parameters to the endEvent() method.

TODO: Still need to find out why a tiny amount of counts is missing when
filtering a larger listfile.