bugfixes and cleanup for the mvlc readout code
Output zip file is now immediately closed on error.

Attempted to move MCST sequence from mesytec-mvlc to the mvme layer to get
better logging (VMEScript vs basic mesytec-mvlc commands). The attempt
failed because the lower level readout worker is then missing the errors
from the mcst sequence and thus does not stop itself. Error handling and
stopping of the readout worker has to be done in the calling code.