trigger io sim: fix strobed LUT output for when the input signal is 0
The previous version did output the inverted input signal after the strobe
pulse was over. This is not correct for 0 input signals: in this case the
strobe does not produce any change in the output at all. This is fixed in
this commit by appending Edge::Falling to the output trace when the strobe
pulse is over.