Revert "core: when a new vme config is loaded reconfigure the analysis"
This reverts commit 4bfb86282b4071ae7fa8577c852768eea216dd5b.

This change was released with mvme-1.10.

Reverted because of a bad side effect with the ExportSink operator:
calling Analysis::begin_run() calls a2_begin_run() which in turn calls
the begin_run() functions of all operators. In export_sink_begin_run()
the output files are recreated, overwriting existing files. This really
must only be called when a daq run or replay is started.

TODO: find another solution to the stale VMEConfig pointer in the
Analysis structure.