analysis: fix errors in the Difference operator with unequal input sizes
Somehow analysis files exist in the wild where the two inputs of the
difference operator have different sizes. In debug builds this asserted,
in release builds it most likely did overwrite some memory by accident.

Now the minimum input size is used for the output size and when performing
the difference calculations.