mdpp_sampling: cleanup and work on a plot ui
TODO:

Record per channel trace history. Add empty entries for missing
channels, so that the same history index corresponds to the same
event number for all channel numbers.

Display a single trace in a plot by default. Allow to scroll through
the trace history and to switch channels.

Allow to display all traces of the same channel in a single window.
Allow to display the traces of all channels in a single event in a single
window.