analysis: better handling of errors when building data sources
Example test case is to add a MdppSampleDecoder to an non-MDPP module.
No more assertions or uncaught exceptions but the error is still not
clearly visible in the ui: neither the data source nor any attached
operators get an error marking. No error is logged, it's all swallowed in
the a2 adapter layer.