waveforms: can now toggle between "latest" and "event snapshot" modes
"latest" samples the latest trace from each channel, prepending it to the
UIs trace history.

"event snapshot" uses a copy of the trace history of the analysis and
display those interpolated traces. In this case the traces of the same
index in the trace history come from the same readout event.