work on a possible tcp stream server implementation
Goal is to have way to monitor raw readout data from other processes.
The server implement IStreamBufferConsumer and this can be attached
to the analysis side StreamWorker instance. This way clients won't
slow the DAQ as the input data from the server comes from the non-blocking
queue between readout and analysis.