analysis: deprecate the ConditionFilter operator
This was added way before the condition system was a thing. It checks
the _validity_ of each input element to determine the condition outcome.
The newer Expression, Interval and Polygon conditions always produce
valid outputs but with values 0/1 for false/true. This means the
ConditionFilter is not compatible with the newer condition system.