templates: mdpp32_scp: change readout_num_events from +2 to +1
This is to make the readout work with sampling enabled and more than ~350
samples. In this case the module buffer can only hold ~2.5 full events if
all channels respond. With the old setting this lead to a situation
where the module had data but never pulled its IRQ due to not reaching
the threshold.

This will break blindly created multievent setups! They where broken before
so at least no additional breakage is added :)