mdpp_sam: rewrite the decoder monitor ui for less lag
Event at >= 1s refresh interval the monitor widget somehow added a ton of
input lag to the whole application. Eventually editing a vme script in
another window would not take key presses at all anymore. Curiously mouse
interactions were still possible.

The fix is to use the same approach as in the MVLC DSO widget: run the
code and text formatting operations in a separate thread and pass the
result back to the ui. Then update the various text browsers in the widget.

Times measured in debug builds:
- pre change : >= 250 ms per update
- post change: ~10 ms to run the decoder and produce the text
               ~30 ms to update the Qt ui