h2d: change the way projections are done to fix zooming/resolution issues
Prior to this change the projection was done over the zoomed in area of the
source histograms. This resulted in non-power-of-2 bin counts which in turn
made the resolution combo defunct.

Now the projection is done over the full physical width/height or height
and the zoom is applied to the projected histogram.

Play with this more to see if it actually improves things.