remove lmd2mvme and gsi-mbs-api: does not compile with gcc-15
The MBS code does use old pre-ansi standard function prototypes which
gcc-15 does not accept anymore. Porting over is mostly just changing the
prototypes from new to old style and making sure everything is consistent.
But it somehow does not seem worth it at the moment. I'll revive it the
next time we work with GSI and need the tool back.