h2d: more projection fixes/updates
Now always uses the visible resolution of the source for the physical
resolution of the projection. No more fiddling with res reduction in the
projection needed.
Auto zoom enabled again.

Note: when zooming into sub-bin ranges the counts of the projections may
differ from the counts displayed in the h2d widget. This is mostly a
display and stats collection issue, not a bug in the algorithm.
TODO: Do another pass to pick bins in a way that the displayed counts in
the source match the counts in the projections.